+++ src/poly_aniso/dipol_exchange.f
@@ -13,14 +13,14 @@ c this Subroutine computes the dipolar coupling between the two moments
       Implicit None
       Integer, parameter            :: wp=SELECTED_REAL_KIND(p=15,r=307)
       Integer, intent(in)           :: N1, N2
-      Real(kind=wp), intent(in)     :: vec(3), dist
-      Complex(kind=wp), intent(in)  :: M1(3,N1,N1)
-      Complex(kind=wp), intent(in)  :: M2(3,N2,N2)
-      Complex(kind=wp), intent(out) :: HDIP(N1,N1,N2,N2)
+      Real(kind=8), intent(in)     :: vec(3), dist
+      Complex(kind=8), intent(in)  :: M1(3,N1,N1)
+      Complex(kind=8), intent(in)  :: M2(3,N2,N2)
+      Complex(kind=8), intent(out) :: HDIP(N1,N1,N2,N2)
 c local variables
       Integer          :: m,i1,j1,i2,j2
-      Real(kind=wp)    :: MB2
-      Complex(kind=wp) :: p2a, p2b, p1, HL, d3, mb2c, threeC, vec1(3)
+      Real(kind=8)    :: MB2
+      Complex(kind=8) :: p2a, p2b, p1, HL, d3, mb2c, threeC, vec1(3)
 
       Call qEnter('PA_Dipol_Exchange')
